requirements.txt
###############################################################################
#
# IMPORTANT:
#
# If runtime or build time dependencies are changed in this file, the committer
# *must* ensure the DEB and RPM package maintainers are informed as soon as
# possible.
#
###############################################################################

Flask==3.0.1
Flask-Login==0.*
Flask-Mail==0.*
Flask-Migrate==4.*
Flask-SQLAlchemy==3.1.*
Flask-WTF==1.2.*
Flask-Compress==1.*
Flask-Paranoid==0.*
Flask-Babel==4.0.*
Flask-Security-Too==5.5.*
Flask-SocketIO==5.4.*
WTForms==3.1.*
passlib==1.*
pytz==2024.*
speaklater3==1.*
sqlparse==0.*
psutil==6.0.*
psycopg[binary]==3.2.3
python-dateutil==2.*
SQLAlchemy==2.*
bcrypt==4.2.*
cryptography==43.0.*
sshtunnel==0.*
ldap3==2.*
gssapi==1.8.*
eventlet==0.37.0
user-agents==2.2.0
Authlib==1.3.*
pyotp==2.*
qrcode[pil]==8.*
boto3==1.35.*
urllib3==1.26.*
keyring==25.*
Werkzeug==3.0.*
typer[all]==0.12.*
setuptools==75.*; python_version >= '3.12'
jsonformatter~=0.3.2
libgravatar==1.0.*